refactor create methods to allow for nesting mestages one level deep
Nested messages are assigned a parent id, which is the id of the comment they will be nested under. Replying to a nested
comment will assign the reply message's parent id value as that of the comment being replied to. The assigning of a parent id is reciprocal, with the parent receiving the childs id in its comments storage. ie. If someone replies to a top level comment (one level of nesting, or replying to a top-level comment), and then I reply
to that message, then they will both be nested at the same depth (one level).

Nesting depth can be deepened by assigning an amount of parent id's equal to the amount of desired nesting.