remove subdirectory that holds user's image type
removes creation of subdirectories (profile, posts, banner), as they are unnecessary now. Given that
the directories are temporary (they are deleted after the image is uploaded to S3), and that the user can only upload one kind of image at a time, it is no longer necessary to keep the image types separated.